requirements.txt
langchain
langchain_community
langchain_huggingface
faiss-cpu
pypdf
huggingface_hub
flask
streamlit
python-dotenv
sentence-transformers